requirements.txt
python-telegram-bot==20.3
requests==2.31.0
psycopg2-binary==2.9.3
python-telegram-bot==20.1
